refactor: Remove unnecessary files and update about page
This commit removes an unnecessary workflow file and a resume file in Simplified Chinese. It also updates the content of the about page to provide more information about the developer, including contact details and links to GitHub and the university. The changes improve the clarity and relevance of the project. YoungZM339/blog@dfccf2fc8fc2df55915cabb27875e4921fba411a